update dependency on p.json to remove error on expo. removed unused component. organized components into pages directory. tested to make sure all imports are properly being called on the path
