--- components/Map.jsx
@@ -1,13 +1,0 @@
-// import React from 'react';
-// import { MapView } from 'expo';
-
-
-// const Map = () => {
-
-
-//     return (
-//         <MapView />
-//     )
-// }
-
-// export default Map;
