+++ components/pages/Home.jsx
@@ -5,10 +5,10 @@ import { ScrollView, TouchableOpacity } from "react-native-gesture-handler";
 import Button from "apsl-react-native-button";
 import axios from 'axios';
 
-import Logo from "./Logo";
+import Logo from "../Logo";
 
-import Styles from "../css/styles";
-import Separator from "./Separator";
+import Styles from "../../css/styles";
+import Separator from "../Separator";
 
 class Home extends Component {
   //Initializing state to capture input
